requirements.txt
fastapi
pydantic>=2.0
pydantic-ai
google-genai
pydantic-settings
python-dotenv

lancedb
numpy
pandas
requests
httpx
python-dotenv
rich
azure-functions
uvicorn
pypdf
streamlit
ipykernel
pydantic-ai

